feat: Add Create function for githubipallowlist_ip_allow_list_entry resource (#5)
This change adds a `githubipallowlist_ip_allow_list_entry` resource with a limited implementation. It implements one of the functions managing `githubipallowlist_ip_allow_list_entry`, Create, on an organization level. The resource still lacks Read, Update and Delete functions, so it's not fully functional. Intention of this change is to provide a chunk of code that's manageable for a reviewer together that can actually be run with a real results (changes to the allow list).

To validate changes run the `simple` example from `examples` directory:

1. Follow [Building The Provider section from README](https://github.com/form3tech-oss/terraform-provider-githubipallowlist/blob/5f6a0e2f6bafc5c041256b8fd4512eb8a0b7e936/README.md#building-the-provider)
2. Generate a Personal Access Token (classic) with an `admin:org` scope.
3. Run the following script filling missing variables:

  ```shell
  cd $PROJECT_ROOT/examples/simple
  export GITHUB_TOKEN=$GENERATED_TOKEN
  export GITHUB_ORGANIZATION=$YOUR_ORGANIZATION_NAME
  terraform plan
  terraform apply
  ```

4. Manually delete IP `1.2.3.4/32` allow list entry from your organization.