feat: Add optional read cache
GetOrganizationIPAllowListEntries executes multiple calls to GitHub GraphQL API to read all IP allow-list entries, i.e. it reads all pages. Since GetOrganizationIPAllowListEntries is called in Read function of githubipallowlist_ip_allow_list_entry resource, it's called once for an every resource instance. As a result, the provider fetches the same data over and over again.

To reduce these redundant calls we're introducing an optional simple read cache for GetOrganizationIPAllowListEntries function that's enabled by default.
The implementation uses a simple synchronized map without any eviction policies. This results in a single batch of calls to GitHub GraphQL API for all GetOrganizationIPAllowListEntries calls in the whole lifetime of the client. This behaviour should be enough for purposes of a Terraform provider, as we only refresh the whole state once. All updates return updated values, therefore there's no need for reconciling reads as well.
The change is transparent for the provider's code. Although current implementation is not optimal, we're doing a whole entries slice scan for every entry read, improvements provided by caching are substantial while keeping the implementation simple and testable in unit and integration tests. Further improvements and optimizations are left for future.