feat: Add support for an enterprise allow list
The provider only supports managing GitHub IP allow list entries on an organization level.

This change adds support for managing GitHub IP allow list entries on an enterprise level.

Client's implementation is based on an existing code. The only difference is a function fetching allow list entries as the API differs between organization and enterprise.
Although GetEnterpriseIPAllowListEntries is a new method, the code itself is basically a copy/paste from GetOrganizationIPAllowListEntries with adapted queries. Given there's already a high degree of code reuse in the codebase (or at least I think so) I felt like duplicating these two methods make sense considering they have different purpose and implementations may diverge in the future.
To adapt provider's code to client's changes I introduced a simple dependency injection. This way it's transparent to a resource whether it's created on an enterprise or an organization level.